requirements.txt
python-dotenv
requests
beautifulsoup4
lxml
tenacity
mysql-connector-python
openpyxl
bcrypt
 